requirements.txt
jieba==0.42.1
Pillow==9.2.0
matplotlib==3.5.2
numpy==1.21.5
wordcloud==1.8.2.2